Add dispatcher and pin handler, reorganize codebase
