+++ src/bots/pins/index.js
@@ -1,0 +1,4 @@
+import client from '../../client';
+import handler from './handler';
+
+client.on('pins', (args) => handler(args[0]));
