+++ src/utils/dispatcher.js
@@ -1,0 +1,17 @@
+import client, { prefix, validCmds } from './constants';
+
+export const parseMessage = (msg) => {
+  msg = msg.toString();
+  if (msg.substring(0, 2) === prefix) {
+    const args = msg.substring(2).split(' ');
+    const cmd = args[0];
+    if (validCmds.includes(cmd)) {
+      client.emit('command', cmd, args.splice(1));
+    }
+  }
+};
+
+export const emitCmd = (cmd, args) => {
+  // TODO: check perms
+  client.emit(cmd, args);
+};
